feature: added exit function and 3way-sync example
