--- examples/git-sync-upstream.flowl
@@ -1,19 +1,0 @@
-load "go:git_sync_upstream"
-load "go:http_get"
-load "go:outcome"
-
-var resp 
-co http_get -> resp {
-    "url": "https://api.github.com/repos/skoo87/cofx"
-    "path": "parent.clone_url"
-}
-
-var out
-co git_sync_upstream ->out {
-    "upstream": "$(resp.parent.clone_url)"
-}
-
-co outcome {
-    "upstream url": "$(resp.parent.clone_url)"
-    "auto sync upstream": "$(out.outcome)"
-}
