+++ std/http/http_get/http_get.go
@@ -89,6 +89,7 @@ func Entrypoint(ctx context.Context, bundle spec.EntrypointBundle, args spec.Ent
 			k := paths[i]
 			v := res.String()
 			returns[k] = v
+			fmt.Fprintf(bundle.Resources.Logwriter, "➜ %s: %s\n", k, v)
 		}
 	}
 	return returns, nil
